lambda_btc_predictor/requirements.txt
hopsworks
wheel
pyproject-toml
python-ffmpeg
binance-connector
pandas
numpy
python-dotenv
hsfs
twofish
sagemaker
boto3
tensorflow
boto
requests